pk 데이터 타입 수정

+++ src/main/java/com/project/LMSSU/Repository/WeeksInfoRepository.java
@@ -3,5 +3,5 @@ package com.project.LMSSU.Repository;
 import com.project.LMSSU.domain.WeeksInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WeeksInfoRepository extends JpaRepository<WeeksInfo, Long> {
+public interface WeeksInfoRepository extends JpaRepository<WeeksInfo, Integer> {
 }
